Run server in SSL
Turns out that there is no way to configure ManageIQ to not make
requests against openshift in SSL:

    https://github.com/ManageIQ/manageiq-providers-kubernetes/blob/9257f08/app/models/manageiq/providers/kubernetes/container_manager_mixin.rb#L33-L35

And this has existed for a while now.  Because the client URL is always
built up to use https, the server is expected to serve requests in SSL.
To allow this, configure webrick to use it's SSL/https functionality.